pyproject.toml
[project]
name = "kenp0m-sp0rts-analyzer"
version = "0.1.0"
description = "Advanced NCAA Division I Men's Basketball analytics using KenPom data for research and analysis"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Andy O'Malley" }
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]
keywords = ["kenpom", "sports-analytics", "basketball", "ncaa", "betting", "statistics"]
dependencies = [
    "kenpompy>=0.5",
    "httpx>=0.27",
    "pandas>=2.0",
    "numpy>=1.26",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "scikit-learn>=1.5",
    "pyarrow>=22.0.0",
    "tweepy>=4.16.0",
    "playwright>=1.57.0",
    "openpyxl>=3.1.5",
    "xgboost>=3.1.2",
    "beautifulsoup4>=4.14.3",
    "selenium>=4.39.0",
    "requests>=2.32.5",
    "webdriver-manager>=4.0.2",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40",
    "playwright-stealth>=1.0.6",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "pytest-playwright>=0.4",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.0",
]
all = [
    "kenp0m-sp0rts-analyzer[browser,mcp,dev]",
]

[project.urls]
Homepage = "https://github.com/omalleyandy/kenp0m-sp0rts-analyzer"
Repository = "https://github.com/omalleyandy/kenp0m-sp0rts-analyzer"
Issues = "https://github.com/omalleyandy/kenp0m-sp0rts-analyzer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kenp0m_sp0rts_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 79  # Strict PEP 8 compliance
src = ["src", "tests", "."]  # Include root for analyze_todays_games.py

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
]
ignore = [
    # E501 now enforced - line length checked by formatter at 79 chars
]

[tool.ruff.lint.isort]
known-first-party = ["kenp0m_sp0rts_analyzer"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "anyio>=4.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
